Adopt SocialBundle as basis for ForumBundle
This commit reengineers the ForumBundle so that it has a dependency on
the rhapsody-project/social-bundle repository. The SocialBundle is the
foundation for all of the Rhapsody social contexts (e.g. blog, forum,
etc.). With this change set the ForumBundle is slimmed down immensely
and will benefit from centralized improvements to the social framework.