Improve forum bundle integration with social bundle
This (large) commit adds a lot of functionality around supporting the
concept of forums as an extension to the Rhapsody SocialBundle. Since
the forum bundle became an extension of the social bundle a lot of
changes have been necessary. This commit contains the following changes
along with others, not listed.

- Break up Twig templates into root and partials so that overriding a
  template is easier.
- Remove the Form/Factory namespace.
- Update controllers to work with SocialBundle delegates (leave forum-
  specific delegates in place for a little while still).
- Update managers and repository interactions.
- Update styles.